requirements.txt
pytest~=8.3.5
selenium~=4.30.0
webdriver-manager~=4.0.2
faker
pytest-html
pandas
pytest-rerunfailures
datadog
pytest-xdist
python-dotenv~=1.1.0
allure-python-commons~=2.14.0
flake8
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
